# api-docsFINAL.yaml
openapi: 3.1.0
info:
  title: Virtual Plant Collection API
  version: 1.0.0
  summary: REST API for mobile, web, and desktop apps in the virtual plant collection system.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /users/register:
    post:
      tags: [users]
      summary: Register a new user
      operationId: register
      security: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                value:
                  username: jan_kowalski
                  email: jan@example.com
                  password: StrongPassw0rd!
      responses:
        '200':
          description: User registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/login:
    post:
      tags: [users]
      summary: Login and get JWT
      operationId: login
      security: []
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          description: User is blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /users/me:
    get:
      tags: [users]
      summary: Get current user
      operationId: me
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Current user data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [users]
      summary: Update current user
      operationId: updateMe
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Username already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: Preferred language for messages.
      schema:
        type: string
        enum: [pl-PL, en-US]
        default: pl-PL
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /users/register
        details:
          type: array
          items:
            type: string
...
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          format: password
    RegisterResponse:
      type: object
      required: [id, username, email]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          description: Username or email.
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [id, username, email, token]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        token:
          type: string
          description: JWT access token.
        refreshToken:
          type: string
          description: JWT refresh token.
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
...
    UserResponse:
      type: object
      required: [id, username, email, role, blocked, createdAt, stats]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        blocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        preferredLanguage:
          type: string
          enum: [pl, en]
        bio:
          type: string
        stats:
          $ref: '#/components/schemas/UserStats'
    UserStats:
      type: object
      required: [totalCollections, totalPlants, uniqueSpecies, friendsCount]
      properties:
        totalCollections:
          type: integer
        totalPlants:
          type: integer
        uniqueSpecies:
          type: integer
        friendsCount:
          type: integer
        publicCollections:
          type: integer
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        bio:
          type: string
          maxLength: 1000
        preferredLanguage:
          type: string
          enum: [pl, en]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT